Fix issue when piping the same stream to both the cache and the client
Piping the same `ReadableStream` to both the cache and the client didn't
work in some cases, resulting in the following error being thrown:

	Error: You cannot pipe after data has been emitted from the response.

The solution was to immediately pipe the download stream to a
`PassThrough` stream. Then, we were able to pipe the `PassThrough`
stream to multiple `WritableStreams` without any issue.

The solution was found in the following StackOverflow answer:

http://stackoverflow.com/questions/19553837/node-js-piping-the-same-stream-into-multiple-writable-targets