Pipe stream to a PassThrough before returning to client
If the client tries to read data from the returned stream after a
slight delay, nothing will be read, since data will be already piped to
the cache location.

The solution is to instantly pipe to a PassThrough stream at the same
time that piping to the cache location and return this new stream
instead.